Project 2: Business Analytics
In this project, we analyzed a large spreadsheet of raw user data that captured every user interaction with our website.  Using this data, we built a conversion funnel, we split our users into acquisition cohorts based on the month of their first purchase, and we calculated the retention rates for each of these cohorts.