Final TripleTen Project
In this project, I joined two large datasets that contained all the user and order data for a global restaurant service called Zomato.  I was asked to analyze this company's data with respect to the types of customers who use this service.  Using Power BI, I created an interactive dashboard that shows which groups of customers spend the most money and also what times are most popular for particular groups of customers.  The "Description of Dashboard with Results and Findings.docx" explains how to use the dashboard and read the graphs, and it also contains results and conclusions to help the company Zomato increase revenue.