Project 3: Data Visualization with Tableau
In this project, I was given a large dataset of a national superstore to analyze customer orders and returns.  I used Tableau to generate several different charts and graphs to answer three main questions.  First, I generated several different visualizations to identify the biggest areas of profit and loss for our superstore.  Second, we created many visualizations to find the 3 best combinations of the month of the year and state in which to advertise in which would have the largest return on investment for our superstore.  Our final big question I wanted to answer had to do with which items were being returned to the store.  To answer this question, I need to join the Returns table to our original table and also create a couple of new calculated columns.  Finally, with this data, I was able to create some visualizations to answer the following questions: which products have the highest return rate, and which customers have the highest return rate.