Project 4: Storytelling with Data
In this project, we were given a large Excel dataset about orders and returns for a national superstore.  My first task was to create many different plots and graphs to analyze and find the root causes of returns.  Then I made a dashboard with user-interactive filters and several graphs in order to examine which stores and customers had the highest returns and find ways to reduce the number of returns.  Next, I created a Tableau story to fully explain all my results and conclusions. I even made a screen recording to explain to new users how to operate my dashboard filters and how I drew my conclusions.